requirements.txt
azure-core>=1.29.0
azure-identity>=1.15.0
azure-search-documents>=11.5.1
matplotlib>=3.8.0
openai>=1.40.0
openpyxl>=3.1.0
pandas>=2.1.0
plotly>=5.18.0
pyautogen>=0.2.0
pyodbc>=5.0.1
pyyaml>=6.0.1
seaborn>=0.13.0
streamlit>=1.31.0
